pyproject.toml
[project]
name = "azure-utils-mcp"
version = "0.1.0"
description = "MCP server for Azure Service Bus, Cosmos DB, and PIM role management"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0",
    "azure-cosmos>=4.7",
    "azure-identity>=1.17",
    "azure-mgmt-cosmosdb>=9.0",
    "azure-mgmt-servicebus>=8.0",
    "azure-mgmt-authorization>=4.0",
    "azure-mgmt-subscription>=3.1",
    "azure-servicebus>=7.12",
]

[project.urls]
Homepage = "https://github.com/BrianDeacon/azure-utils-mcp"
Repository = "https://github.com/BrianDeacon/azure-utils-mcp"
Issues = "https://github.com/BrianDeacon/azure-utils-mcp/issues"

[project.scripts]
azure-utils-mcp = "azure_utils_mcp.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_utils_mcp"]
